don't use strnstr(line,…,strlen(line))
That does nothing: strnstr only makes code any more secure than strstr if the length of the string to look through is known. If you have to use strlen, you might as well just use strstr, which does the same.

but unlike strnstr, strstr *IS* POSIX and not BSD-only / libbsd.

This broke buildability on a modern GCC system.

Signed-off-by: Marcus Müller <marcus@hostalia.de>